[2단계 - 상품 관리 기능] 비버(전인표) 미션 제출합니다 (#293)
* refactor: 레이어드아키텍쳐 적용(dto 생성)

* refactor: 서비스클래스 네이밍 수정

* refactor: 패키지 이동

* feat: 데이터 삽입, 테이블 생성

* feat: 사용자 설정페이지 구현

* refactor: 상품저장 기능 네이밍 수정

* feat: 사용자의 검증 로직 생성

* refactor: 출력형식의 맞게 수정

* refactor: service단에서 넘어오는 Dto삭제 및 입력 타입 변경

* feat: 장바구니 기능 구현

* feat: 사용자의 id의 따른 상품 검색 기능 추가

* refactor: 테스트코드 수정

* refactor: api수정

* refactor: 가독성을 위한 계행

* refactor: Transactional추가

* refactor: cart 테스트 작성

* test: /cart get요청 테스트

* test: /cart post요청 테스트

* test: 테스트 샛팅

* refactor: 테스트 메소드 수정

* refactor: 사용하지 않는 변수 삭제

* refactor: 웹으로 부터 응답 수정

* feat: 상품 id를 requestBody로 요청

* refactor: 웹 api 수정

* refactor: test 네이밍 수정

* refactor: dto패키지 분리

* feat: 웹으로 보내는 dto 생성

* refactor: 기능의 따른 service클래스 분리

* refactor: CartItemDeleteRequest 삭제

* refactor: 기능의 따른 service클래스 분리

* refactor: 테이블 네이밍 수정

* docs: 기능목록 업데이트

* refactor: MemberService 패키지 이동

* feat: 검증 어노테이션 추가

* refactor: Transactional위치 변경

* refactor: 변수명 변경

* feat: fixture 적용

* refactor: MemberIdRequest 네이밍 수정

---------

+++ src/main/java/cart/dao/ProductInMemoryDao.java
@@ -11,8 +11,8 @@ import java.util.concurrent.atomic.AtomicInteger;
 @Repository
 public class ProductInMemoryDao implements ProductDao {
 
-    private final List<ProductEntity> products; // 넣는 순서는 상관없으므로 동시성 문제가 없다.
-    private final AtomicInteger sequence = new AtomicInteger(0); // 동시성 문제 해결
+    private final List<ProductEntity> products;
+    private final AtomicInteger sequence = new AtomicInteger(0);
 
     public ProductInMemoryDao() {
         this.products = new ArrayList<>();
@@ -20,7 +20,7 @@ public class ProductInMemoryDao implements ProductDao {
 
     @Override
     public Integer insert(final ProductEntity productEntity) {
-        final int id = sequence.incrementAndGet();
+        final Long id = (long) sequence.incrementAndGet();
         ProductEntity productEntityWithId = new ProductEntity(
                 id,
                 productEntity.getName(),
@@ -29,7 +29,7 @@ public class ProductInMemoryDao implements ProductDao {
         );
 
         products.add(productEntityWithId);
-        return id;
+        return Math.toIntExact(id);
     }
 
     @Override
@@ -39,7 +39,7 @@ public class ProductInMemoryDao implements ProductDao {
     }
 
     @Override
-    public void deleteById(final Integer id) {
+    public void deleteById(final Long id) {
         Optional<ProductEntity> byId = findById(id);
         if (byId.isEmpty()) {
             throw new IllegalArgumentException("아이디를 찾을 수 없습니다.");
@@ -48,7 +48,7 @@ public class ProductInMemoryDao implements ProductDao {
     }
 
     @Override
-    public Optional<ProductEntity> findById(final Integer id) {
+    public Optional<ProductEntity> findById(final Long id) {
         return products.stream()
                 .filter(product -> product.getId().equals(id))
                 .findFirst();
